Factor tt::is_integer into its own file
