+++ src/DataStructures/Index.hpp
@@ -17,6 +17,7 @@
 #include "Utilities/MakeArray.hpp"
 #include "Utilities/Requires.hpp"
 #include "Utilities/TypeTraits.hpp"  // IWYU pragma: keep
+#include "Utilities/TypeTraits/IsInteger.hpp"
 
 namespace PUP {
 class er;
