+++ src/Domain/Mesh.hpp
@@ -13,6 +13,7 @@
 #include "Utilities/Gsl.hpp"
 #include "Utilities/Requires.hpp"
 #include "Utilities/TypeTraits.hpp"  // IWYU pragma: keep
+#include "Utilities/TypeTraits/IsInteger.hpp"
 
 /// \cond
 namespace PUP {
